requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Cloud Platform
google-cloud-bigquery==3.13.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0